pxlTextGenerator at working status
Bug fixes and rotation is fully functional (if it wasn't in the last update)
Minor bugs still exist but don't crash the program.

Some bugs exist-
_Needing to load a character from the viewer before any edited data is stored to existing characters.
_Needing to load a text base image prior to loading exported images.  Adding extra steps before producing images.

ToDo-
Reimpliment the lazy loading system to the character viewer list.  Loading all 340+ characters is slow for an initial load when it just needs to be the database file prior to using the characters in pages or test bed viewing.